Rebranded to only manage AO auth.
Renamed plugin to auth_association_online:
 * Changed logging calls
 * Renamed functions
 * Removed doc comments referring to google
 * _Retained_ all copyright and license notices from original author

Removed travis config, as we don't use travis.

Removed test files, as we will not be running automated tests against this, and
we don't want to have to make sure that the tests written for google/linked-in/etc
work with the new "AO" plugin.

Linked users based on additional "associationOnlineId" field, rather than email.